% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_clustersRegulome}
\alias{create_clustersRegulome}
\title{Create clustersRegulome Object}
\usage{
create_clustersRegulome(coordinates, cluster.type, col = "dark green", genome,
  path = "~/data/IsletRegulome/Rdata/")
}
\arguments{
\item{coordinates}{GRanges object with coordinates you want to plot.}

\item{cluster.type}{Name of the cluster to plot. The value can be: enhancerClusters, enhancerHubs, stretchEnhancers, superEnhancers or "" (no clusters).}

\item{col}{Color for plotting clusters. Default: "dark green".}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}
}
\description{
This function loads the selected cluster classification, selects the clusters present in \code{coordinates}
and adds the needed parameters for plotting.
}
