% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_contactsRegulome}
\alias{create_contactsRegulome}
\title{Create contactsRegulome Object}
\usage{
create_contactsRegulome(coordinates, contacts.type, col = c(`0` = "grey", `3`
  = "blue", `5` = "dark orange"), genome, smooth = 5,
  path = "~/data/IsletRegulome/Rdata/")
}
\arguments{
\item{coordinates}{GRanges object with coordinates you want to plot.}

\item{contacts.type}{ID for the bait you want to plot. See correspondance in \url{http://gattaca.imppc.org/genome_browser/lplab/IsletRegulome/Rdata/hg19/baitID_and_name_virtual4C.rda}}

\item{col}{Colors for plotting virtual 4C data. Defaults to c("grey", "blue", "dark orange").}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{smooth}{Arbitrary number for smoothing contact signal. Default: 5.}
}
\description{
This function uses virtual 4C data (3 bigWig files, scores 0, 3 and 5) to generate an object containing
the data present in \code{coordinates} and all the elements needed for its plotting method.
}
