% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_genesRegulome}
\alias{create_genesRegulome}
\title{Create genesRegulome Object}
\usage{
create_genesRegulome(coordinates, col = c(gene = "dark grey", lnc = "black"),
  showLongestTranscript = TRUE, genome,
  path = "~/data/IsletRegulome/Rdata/")
}
\arguments{
\item{coordinates}{GRanges object with coordinates you want to plot.}

\item{col}{Color for the genes.}

\item{showLongestTranscript}{When plotting gene data, set to TRUE (default) if you want to reduce the number of transcripts by only plotting the longest transcript per gene. If set to FALSE, will plot all the transcripts.}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}
}
\description{
This funciton loads all genes in chr, selects those transcripts present in \code{coordinates}
and adds the needed parameters for plotting.
}
