% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_mapsRegulome}
\alias{create_mapsRegulome}
\title{Create mapsRegulome Object}
\usage{
create_mapsRegulome(coordinates, maps.type, genome,
  path = "~/data/IsletRegulome/Rdata/")
}
\arguments{
\item{coordinates}{GRanges object with coordinates you want to plot.}

\item{maps.type}{Name of the chromatin maps to plot. The value can be: chromatinClasses, chromatinClassesReduced, chromatinStates, openChromatinClasses, progenitors or "" (default, no map).}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}
}
\description{
This function loads chromatin maps, obtains the classes present in \code{coordinates} and adds
the needed parameters for plotting.
}
