% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_snpsRegulome}
\alias{create_snpsRegulome}
\title{Create snpsRegulome Object}
\usage{
create_snpsRegulome(coordinates, snps_dataset = "",
  snps_col = "dark red", genome = "hg19", top = 3,
  snp_scaling = NULL, path = "~/data/IRB/")
}
\arguments{
\item{coordinates}{Either a GRanges object, data.frame or character string (chr11:17227951-17589050) indicating
the coordinates for the region the user wants to plot.}

\item{snps_dataset}{GWAS SNPs dataset name to use for the analysis. The value can be: diagram, magic,
70KforT2D or "".
(default, no SNPs will be plotted).}

\item{snps_col}{Color for the GWAS SNPs. Default: "dark red"}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{path}{Path containing the genomes folder (for example "hg19").
Default: "http://gattaca.imppc.org/genome_browser/lplab/IsletRegulome/Rdata/"}
}
\value{
An object of class \code{snpsRegulome}, which is basically a list with the following
        attributes:
        \itemize{
            \item{\strong{name}: Name of the selected SNP dataset.}
            \item{\strong{col}: Color for plotting the SNPs.}
            \item{\strong{genome}: Genome build.}
            \item{\strong{coordinates}: Selected coordinates to plot as a \code{GRanges} object.}
            \item{\strong{value}: \code{GRanges} object containing all SNP values in \code{coordinates}.}
            \item{\strong{moreArgs}: Additional arguments, which are:
                  \itemize{
                      \item{\strong{top}: contains the
                            number of SNPs with lower P-values whose name will be plot}
                      \item{\strong{scaleTo}: parameter that includes a scaling parameter in case
                            virtual4C is also plot along SNPs (it is the maximum virtual4C value and
                            if no virtual4C is plot, is the maximum SNP -log10 p-value).}
                      }
                 }
            }
}
\description{
This function loads the GWAS information for SNPs from the specified dataset, selects those
present in \code{coordinates} and adds all the elements needed for plotting.
}
