% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobandColor.R
\docType{data}
\name{cytobandColor}
\alias{cytobandColor}
\title{Cytoband color}
\format{Named character vector.}
\usage{
cytobandColor
}
\description{
Dataset containing all colors corresponding to the cytoband named in ideoCyto.
}
\keyword{datasets}
