% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIRB.R
\name{downloadIRB}
\alias{downloadIRB}
\title{Download IRB datasets}
\usage{
downloadIRB(output_dir,
  file_dir = "http://gattaca.imppc.org/genome_browser/lplab/IRB_database.tar.gz")
}
\arguments{
\item{output_dir}{Output directory for the datasets.}

\item{file_dir}{Path to the compressed IRB database file.}
}
\description{
Downloads the required IRB datasets.
}
