% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegulomeObjects.R
\name{emptyPlot}
\alias{emptyPlot}
\title{Generate emtpy plot}
\usage{
emptyPlot(coordinates, y)
}
\arguments{
\item{coordinates}{GRanges object with coordinates you want to plot.}

\item{y}{Value for the empty point in Y axis. Usually set to either 0 or 1.}
}
\description{
In those cases where there's no output, it generates an empty plot.
}
