% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideoCyto.R
\docType{data}
\name{ideoCyto}
\alias{ideoCyto}
\title{Cytoband Ideogram}
\format{List of \code{GRanges} objects, each element corresponding to a genome and the ranges
represented show different cytobands for each chromosome with their stain names.}
\usage{
ideoCyto
}
\description{
Dataset containing all cytobands for different genomes.
}
\keyword{datasets}
