% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRB.R
\docType{data}
\name{IRB}
\alias{IRB}
\title{IRB}
\format{List containing each type of dataset with the available dataset names.}
\usage{
IRB
}
\description{
Dataset containing all available dataset names to use for plotting.
}
\keyword{datasets}
