% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegulomeObjects.R
\name{plot}
\alias{plot}
\title{Generic plotting methof for Regulome data}
\usage{
plot(regulomeObject)
}
\arguments{
\item{regulomeObject}{An object created by \code{create_xxRegulome}, depending on the data you want to show.}
}
\description{
This is a generic method that depending on the class of the passed object will plot according
to the regulome Plot.
}
