% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIdeogramRegulome.R
\name{plotIdeogramRegulome}
\alias{plotIdeogramRegulome}
\title{Plot Ideogram for Regulome Plot}
\usage{
plotIdeogramRegulome(coordinates, genome)
}
\arguments{
\item{coordinates}{GRanges object with coordinates you want to plot.}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}
}
\description{
This function creates an ideogram representation of the chromosome in \code{coordinates}.
}
