% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegulome.R
\name{plotRegulome}
\alias{plotRegulome}
\title{Plot Regulome Data}
\usage{
plotRegulome(coordinates, snps_dataset = "", snps_col = "dark red",
  contacts_dataset = "", contacts_col = c(`0` = "grey", `3` = "blue",
  `5` = "dark orange"), maps_dataset = "", cluster_dataset = "",
  cluster_col = "dark green", tfs_dataset = "",
  tfs_col = "dark blue", genes_col = c(gene = "dark grey", spec =
  "darkorchid3", lnc = "black"), showLongestTranscript = TRUE,
  genome = "hg19", path = "~/data/IRB/")
}
\arguments{
\item{coordinates}{Either a GRanges object, data.frame or character string (chr11:17227951-17589050) indicating
the coordinates for the region the user wants to plot.}

\item{snps_dataset}{GWAS SNPs dataset name to use for the analysis. The value can be: diagram, magic,
70KforT2D or "".
(default, no SNPs will be plotted).}

\item{snps_col}{Color for the GWAS SNPs. Default: "dark red"}

\item{contacts_dataset}{BaitID or bait gene name for the virtual 4C data you want to plot.}

\item{maps_dataset}{Name of the chromatin maps to plot. The value can be: chromatinClasses,
chromatinClassesReduced, chromatinStates, openChromatinClasses, progenitors or "" (default, no map).}

\item{cluster_dataset}{Name of the cluster to plot. The value can be: enhancerClusters, enhancerHubs,
stretchEnhancers, superEnhancers or "" (no clusters).}

\item{cluster_col}{Color for the clusters. Default: "dark green"}

\item{tfs_dataset}{Name of the TF dataset to plot. The value can be: adult, progenitors, structure or ""
(don't plot TFs).}

\item{tfs_col}{Color for the tfs circles border.}

\item{genes_col}{Named character vector with the colors for each type of feture plotted in the gene
annotation track.}

\item{showLongestTranscript}{When plotting gene data, set to TRUE (default) if you want to reduce the
number of transcripts by only plotting the longest transcript per gene. If set to FALSE, will plot all
the transcripts.}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{path}{Path containing the genomes folder (for example "hg19").
Default: "http://gattaca.imppc.org/genome_browser/lplab/IsletRegulome/Rdata/"}
}
\value{
A ggplot2 object that can be plotted or saved with ggsave.
}
\description{
Plot Regulome Data
}
