% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themesBlank.R
\name{scaleXCoordinates}
\alias{scaleXCoordinates}
\title{X axis for genomic coordinates}
\usage{
scaleXCoordinates(chr, ...)
}
\arguments{
\item{chr}{Logical indicating if the title of the axis should be included. Default = FALSE.}

\item{...}{Additional parameters to pass to \link[ggplot2]{scale_x_continuous} function.}
}
\description{
When added to a ggplot, will set the X axis to represent genomic coordintes data, by rounding
X axis labels to Mb and showing the chromosome number as an X axis title.
}
