% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegulomeObjects.R
\name{smoothCoverage}
\alias{smoothCoverage}
\title{Smooth Coverage/Contacts}
\usage{
smoothCoverage(coverage.gr, smooth = 5, coordinates)
}
\arguments{
\item{coverage.gr}{GRanges object with the first mcol containing the score of the contact.}

\item{smooth}{Arbitary number with smoothing value. Default: 5}

\item{coordinates}{GRanges object with coordinates you want to plot.}
}
\description{
Given a GRanges dataset that contains scores in the first mcol, it will smooth the data.
}
