% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themesBlank.R
\name{themeXblank}
\alias{themeXblank}
\title{Blank X axis ggplot2 theme}
\usage{
themeXblank(...)
}
\arguments{
\item{...}{Additional parameters to pass to \link[ggplot2]{theme} function.}
}
\description{
When added to a ggplot, will set all elements in the x axis to blank.
}
