% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themesBlank.R
\name{themeXYblank}
\alias{themeXYblank}
\title{Blank X and Y axis ggplot2 theme}
\usage{
themeXYblank(title = FALSE, ...)
}
\arguments{
\item{title}{Logical indicating if the title of the axis should be included. Default = FALSE.}

\item{...}{Additional parameters to pass to \link[ggplot2]{theme} function.}
}
\description{
When added to a ggplot, will set all elements in the x and y axis to blank.
}
