% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themesBlank.R
\name{themeYblank}
\alias{themeYblank}
\title{Blank Y axis ggplot2 theme}
\usage{
themeYblank(title = FALSE, ...)
}
\arguments{
\item{title}{Logical indicating if the title of the axis should be included. Default = FALSE.}

\item{...}{Additional parameters to pass to \link[ggplot2]{theme} function.}
}
\description{
When added to a ggplot, will set all elements in the y axis to blank.
}
