% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_clustersRegulome}
\alias{create_clustersRegulome}
\title{Create clustersRegulome Object}
\usage{
create_clustersRegulome(coordinates, clusters_dataset,
  cluster_col = "dark green", genome = "hg19", path = "~/data/IRB/")
}
\arguments{
\item{coordinates}{Either a \code{GRanges} object, data.frame or character string
(i.e. \code{chr11:17227951-17589050}) indicating the coordinates for the region the user wants
to plot.}

\item{clusters_dataset}{Name of the enhncer cluster data to plot. The value should be one
of the defined below (column \emph{Dataset}). Defaults to \code{""}, which will produce an
empty plot.
\tabular{lll}{
\strong{Name}                     \tab \strong{Dataset}        \tab \strong{Reference}  \cr
Enhancer Hubs                     \tab enhancerHubs            \tab Miguel-Escalada et al. (2019)\cr
Super Enhancers                   \tab superEnhancers          \tab Miguel-Escalada et al. (2019)\cr
Enhancer Clusters                 \tab enhancerClusters        \tab Pasquali et al. (2014)       \cr
Stretch Enhancers                 \tab stretchEnhancers        \tab Parker et al. (2013)         \cr
COREs                             \tab cores                   \tab K. J. Gaulton et al. (2010)        \cr
}}

\item{cluster_col}{Color for the clusters. Default: "dark green"}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{path}{Path containing the genomes folder (for example "hg19").
Default: "~/data/IRB/"}
}
\value{
An object of class \code{clustersRegulome}, which is basically a list with the following
        attributes:
        \itemize{
            \item{\strong{name}: Name of selected enhancer cluster dataset.}
            \item{\strong{col}: Character vector with the color of the enhancer clusters.}
            \item{\strong{genome}: Genome build.}
            \item{\strong{coordinates}: Selected coordinates to plot as a \code{GRanges} object.}
            \item{\strong{value}: \code{GRanges} object containing coordinates for the enhancer clusters and
                the class of the line that will be drawn.}
            \item{\strong{moreArgs}: Additional arguments, which include:
                \itemize{
                    \item{\strong{lwd}: Information on the linewidth for each class of enhancer cluster
                        feature.}
                    }
                 }
            }
}
\description{
This function loads the selected cluster classification, selects the clusters present in \code{coordinates}
and adds the needed parameters for plotting.
}
\examples{
\dontrun{
clust <- create_clustersRegulome(coordinates="chr20:22461643-22666093",
                                 clusters_dataset="enhancerHubs",
                                 path="~/data/IRB_database/")
}
}
