% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_mapsRegulome}
\alias{create_mapsRegulome}
\title{Create mapsRegulome Object}
\usage{
create_mapsRegulome(coordinates, maps_dataset, genome = "hg19",
  path = "~/data/IRB/")
}
\arguments{
\item{coordinates}{Either a \code{GRanges} object, data.frame or character string
(i.e. \code{chr11:17227951-17589050}) indicating the coordinates for the region the user wants
to plot.}

\item{maps_dataset}{Name of the chromatin maps to plot. The value should be one
of the defined below (column \emph{Dataset}). Defaults to \code{""}, which will produce an
empty plot.
\tabular{lll}{
\strong{Name}                     \tab \strong{Dataset}        \tab \strong{Reference}  \cr
Adult Islets - Chromatin Classes  \tab chromatinClassesReduced \tab Miguel-Escalada et al. (2019)\cr
Adult Islets -  Chromatin Classes \tab openChromatinClasses    \tab Pasquali et al. (2014)       \cr
Pancreatic Progenitors            \tab progenitors             \tab Cebola et al. (2015)         \cr
Adult Islets - Chromatin States   \tab chromatinStates         \tab Parker et al. (2013)         \cr
}}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{path}{Path containing the genomes folder (for example "hg19").
Default: "~/data/IRB/"}
}
\value{
An object of class \code{mapsRegulome}, which is basically a list with the following
        attributes:
        \itemize{
            \item{\strong{name}: Name of selected chromatin map dataset.}
            \item{\strong{col}: Named vector with the color of each type of chromatin class.}
            \item{\strong{genome}: Genome build.}
            \item{\strong{coordinates}: Selected coordinates to plot as a \code{GRanges} object.}
            \item{\strong{value}: \code{GRanges} object containing coordinates and type of chromatin class.}
            \item{\strong{moreArgs}: Additional arguments, which are empty.}
            }
}
\description{
This function loads chromatin maps, obtains the classes present in \code{coordinates} and adds
the needed parameters for plotting.
}
