% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_snpsRegulome}
\alias{create_snpsRegulome}
\title{Create snpsRegulome Object}
\usage{
create_snpsRegulome(coordinates, snps.type, col = "dark red", genome,
  top = 3, maxContacts = NULL, path = "~/data/IsletRegulome/Rdata/")
}
\arguments{
\item{coordinates}{GRanges object with coordinates you want to plot.}

\item{snps.type}{GWAS SNPs dataset name to use for the analysis. The value can be: diagram, magic, 70KforT2D or ""
(default, no SNPs will be plotted).}

\item{col}{Color for the GWAS SNPs. Default: "dark red"}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{top}{Number of most significant SNPs for showing ID. Default: 3.}

\item{maxContacts}{Maximum score for contacts in contactsRegulome object. If no contacts present, should be NULL.}
}
\description{
This function loads the GWAS information for SNPs from the specified dataset, selects those
present in \code{coordinates} and adds all the elements needed for plotting.
}
