% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_snpsRegulome}
\alias{create_snpsRegulome}
\title{Create snpsRegulome Object}
\usage{
create_snpsRegulome(coordinates, snps_dataset = "",
  snps_col = "dark red", genome = "hg19", top = 3,
  snps_scaling = NULL, path = "~/data/IRB/")
}
\arguments{
\item{coordinates}{Either a \code{GRanges} object, data.frame or character string
(i.e. \code{chr11:17227951-17589050}) indicating the coordinates for the region the user wants
to plot.}

\item{snps_dataset}{GWAS SNPs dataset name to use for the analysis. The value should be one
of the defined below (column \emph{Dataset}). Defaults to \code{""}, which will produce an
empty plot.
\tabular{lll}{
\strong{Name}                     \tab \strong{Dataset}        \tab \strong{Reference}  \cr
70KforT2D                         \tab 70KforT2D               \tab S. Bonàs-Guarch et al. (2018)   \cr
Diagram                           \tab diagram                 \tab A. P. Morris et al. (2012)         \cr
Magic                             \tab magic                   \tab R. A. Scott et al. (2012)          \cr
}}

\item{snps_col}{Color for the GWAS SNPs. Default: "dark red"}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{path}{Path containing the genomes folder (for example "hg19").
Default: "~/data/IRB/"}
}
\value{
An object of class \code{snpsRegulome}, which is basically a list with the following
        attributes:
        \itemize{
            \item{\strong{name}: Name of the selected SNP dataset.}
            \item{\strong{col}: Color for plotting the SNPs.}
            \item{\strong{genome}: Genome build.}
            \item{\strong{coordinates}: Selected coordinates to plot as a \code{GRanges} object.}
            \item{\strong{value}: \code{GRanges} object containing all SNP values in \code{coordinates}.}
            \item{\strong{moreArgs}: Additional arguments, which are:
                  \itemize{
                      \item{\strong{top}: contains the
                            number of SNPs with lower P-values whose name will be plot}
                      \item{\strong{scaleTo}: parameter that includes a scaling parameter in case
                            virtual4C is also plot along SNPs (it is the maximum virtual4C value and
                            if no virtual4C is plot, is the maximum SNP -log10 p-value).}
                      }
                 }
            }
}
\description{
This function loads the GWAS information for SNPs from the specified dataset, selects those
present in \code{coordinates} and adds all the elements needed for plotting.
}
