% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_tfsRegulome}
\alias{create_tfsRegulome}
\title{Create tfsRegulome Object}
\usage{
create_tfsRegulome(coordinates, tfs.type, col = "dark blue",
  position.y = 0.5, genome, path = "~/data/IsletRegulome/Rdata/")
}
\arguments{
\item{coordinates}{GRanges object with coordinates you want to plot.}

\item{tfs.type}{Name of the TF dataset to plot. The value can be: adult, progenitors, structure or "" (don't plot TFs).}

\item{col}{Color for the tfs circles border.}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{position}{Position of the center of TFs circle, should be between c(0,1).}
}
\description{
This function load the TFBS data, selects those binding sites present in \code{coordinates} and
adds the needed parameters for plotting.
}
