% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulomeObjects.R
\name{create_tfsRegulome}
\alias{create_tfsRegulome}
\title{Create tfsRegulome Object}
\usage{
create_tfsRegulome(coordinates, tfs_dataset, tfs_col = "dark blue",
  position_y = 0.5, genome = "hg19", path = "~/data/IRB/")
}
\arguments{
\item{coordinates}{Either a \code{GRanges} object, data.frame or character string
(i.e. \code{chr11:17227951-17589050}) indicating the coordinates for the region the user wants
to plot.}

\item{tfs_dataset}{Name of the TF dataset to plot. The value should be one
of the defined below (column \emph{Dataset}). Defaults to \code{""}, which will produce an
empty plot.
\tabular{lll}{
\strong{Name}                     \tab \strong{Dataset}        \tab \strong{Reference}  \cr
Adult Islets – Structural         \tab structure               \tab Miguel-Escalada et al. (2019)\cr
Pancreatic Progenitors            \tab progenitors             \tab Cebola et al. (2015)         \cr
Adult Islets – Tissue-specific    \tab adult                   \tab Pasquali et al. (2014)
}}

\item{tfs_col}{Color for the TFs circle border.}

\item{position_y}{Number from 0 (bottom) to 1 (top) indicating the position at which the circles should
be plotted. Default = 0.5.}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{path}{Path containing the genomes folder (for example "hg19").
Default: "~/data/IRB/"}
}
\value{
An object of class \code{tfsRegulome}, which is basically a list with the following
        attributes:
        \itemize{
            \item{\strong{name}: Name of selected transcription factor dataset.}
            \item{\strong{col}: Character vector with the color of the enhancer clusters.}
            \item{\strong{genome}: Genome build.}
            \item{\strong{coordinates}: Selected coordinates to plot as a \code{GRanges} object.}
            \item{\strong{value}: \code{GRanges} object containing coordinates for the transcription factor binding
                sites and the name of the TF that is binding.}
            \item{\strong{moreArgs}: Additional arguments, which include:
                \itemize{
                    \item{\strong{positionY}: Information on the linewidth for each class of enhancer cluster
                        feature.}
                    }
                 }
            }
}
\description{
This function load the TFBS data, selects those binding sites present in \code{coordinates} and
adds the needed parameters for plotting.
}
