% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegulome.R
\name{plotRegulome}
\alias{plotRegulome}
\title{Plot Regulome Data}
\usage{
plotRegulome(coordinates, snps.type = "", snps.col = "dark red",
  contacts.type = "", maps.type = "", cluster.type = "",
  cluster.col = "dark green", tfs.type = "", tfs.col = "dark blue",
  showLongestTranscript = TRUE, genome = "hg19",
  path = "http://gattaca.imppc.org/genome_browser/lplab/IsletRegulome/Rdata/")
}
\arguments{
\item{coordinates}{Either a GRanges object, data.frame or character string (chr11:17227951-17589050) indicating
the coordinates for the region the user wants to plot.}

\item{snps.type}{GWAS SNPs dataset name to use for the analysis. The value can be: diagram, magic, 70KforT2D or ""
(default, no SNPs will be plotted).}

\item{snps.col}{Color for the GWAS SNPs. Default: "dark red"}

\item{contacts.type}{ID for the bait you want to plot. See correspondance in \url{http://gattaca.imppc.org/genome_browser/lplab/IsletRegulome/Rdata/hg19/baitID_and_name_virtual4C.rda}}

\item{maps.type}{Name of the chromatin maps to plot. The value can be: chromatinClasses, chromatinClassesReduced, chromatinStates, openChromatinClasses, progenitors or "" (default, no map).}

\item{cluster.type}{Name of the cluster to plot. The value can be: enhancerClusters, enhancerHubs, stretchEnhancers, superEnhancers or "" (no clusters).}

\item{cluster.col}{Color for the clusters. Default: "dark green"}

\item{tfs.type}{Name of the TF dataset to plot. The value can be: adult, progenitors, structure or "" (don't plot TFs).}

\item{tfs.col}{Color for the tfs circles border.}

\item{showLongestTranscript}{When plotting gene data, set to TRUE (default) if you want to reduce the number of transcripts by only plotting the longest transcript per gene. If set to FALSE, will plot all the transcripts.}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{path}{Path containing the genomes folder (for example "hg19"). Default: "http://gattaca.imppc.org/genome_browser/lplab/IsletRegulome/Rdata/"}
}
\value{
A ggplot2 object that can be plotted or saved with ggsave.
}
\description{
Plot Regulome Data
}
