% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegulome.R
\name{plotRegulome}
\alias{plotRegulome}
\title{Plot Regulome Data}
\usage{
plotRegulome(coordinates, snps_dataset = "", snps_col = "dark red",
  contacts_dataset = "", contacts_col = c(`0` = "grey", `3` = "blue",
  `5` = "dark orange"), maps_dataset = "", cluster_dataset = "",
  cluster_col = "dark green", tfs_dataset = "",
  tfs_col = "dark blue", genes_col = c(gene = "dark grey", spec =
  "darkorchid3", lnc = "black"), showLongestTranscript = TRUE,
  randomIRB = FALSE, genome = "hg19", path = "~/data/IRB/")
}
\arguments{
\item{coordinates}{Either a \code{GRanges} object, data.frame or character string
(i.e. \code{chr11:17227951-17589050}) indicating the coordinates for the region the user wants
to plot.}

\item{snps_dataset}{GWAS SNPs dataset name to use for the analysis. The value should be one
of the defined below (column \emph{Dataset}). Defaults to \code{""}, which will produce an
empty plot.
\tabular{lll}{
\strong{Name}                     \tab \strong{Dataset}        \tab \strong{Reference}  \cr
70KforT2D                         \tab 70KforT2D               \tab S. Bonàs-Guarch et al. (2018)   \cr
Diagram                           \tab diagram                 \tab A. P. Morris et al. (2012)         \cr
Magic                             \tab magic                   \tab R. A. Scott et al. (2012)          \cr
}}

\item{snps_col}{Color for the GWAS SNPs. Default: "dark red"}

\item{contacts_dataset}{BaitID or bait gene name for the virtual 4C data (Miguel-Escalada
et al. (2019)) to plot.}

\item{maps_dataset}{Name of the chromatin maps to plot. The value should be one
of the defined below (column \emph{Dataset}). Defaults to \code{""}, which will produce an
empty plot.
\tabular{lll}{
\strong{Name}                     \tab \strong{Dataset}        \tab \strong{Reference}  \cr
Adult Islets - Chromatin Classes  \tab chromatinClassesReduced \tab Miguel-Escalada et al. (2019)\cr
Adult Islets -  Chromatin Classes \tab openChromatinClasses    \tab Pasquali et al. (2014)       \cr
Pancreatic Progenitors            \tab progenitors             \tab Cebola et al. (2015)         \cr
Adult Islets - Chromatin States   \tab chromatinStates         \tab Parker et al. (2013)         \cr
}}

\item{cluster_dataset}{Name of the enhncer cluster data to plot. The value should be one
of the defined below (column \emph{Dataset}). Defaults to \code{""}, which will produce an
empty plot.
\tabular{lll}{
\strong{Name}                     \tab \strong{Dataset}        \tab \strong{Reference}  \cr
Enhancer Hubs                     \tab enhancerHubs            \tab Miguel-Escalada et al. (2019)\cr
Super Enhancers                   \tab superEnhancers          \tab Miguel-Escalada et al. (2019)\cr
Enhancer Clusters                 \tab enhancerClusters        \tab Pasquali et al. (2014)       \cr
Stretch Enhancers                 \tab stretchEnhancers        \tab Parker et al. (2013)         \cr
COREs                             \tab cores                   \tab K. J. Gaulton et al. (2010)        \cr
}}

\item{cluster_col}{Color for the clusters. Default: "dark green"}

\item{tfs_dataset}{Name of the TF dataset to plot. The value should be one
of the defined below (column \emph{Dataset}). Defaults to \code{""}, which will produce an
empty plot.
\tabular{lll}{
\strong{Name}                     \tab \strong{Dataset}        \tab \strong{Reference}  \cr
Adult Islets – Structural         \tab structure               \tab Miguel-Escalada et al. (2019)\cr
Pancreatic Progenitors            \tab progenitors             \tab Cebola et al. (2015)         \cr
Adult Islets – Tissue-specific    \tab adult                   \tab Pasquali et al. (2014)
}}

\item{tfs_col}{Color for the TFs circle border.}

\item{genes_col}{Named character vector with the colors for each type of feture plotted in the gene
annotation track.}

\item{showLongestTranscript}{When plotting gene data, set to TRUE (default) if you want to reduce the
number of transcripts by only plotting the longest transcript per gene. If set to FALSE, will plot all
the transcripts.}

\item{randomIRB}{Generate random combinations of IRB datasets.}

\item{genome}{Character string indicating the genome for the coordinates. Default: hg19.}

\item{path}{Path containing the genomes folder (for example "hg19").
Default: "~/data/IRB/"}
}
\value{
A ggplot2 object that can be plotted or saved with ggsave. Recomended device size in
inches is: \code{width=??, height=??}
}
\description{
By providing the coordinates and IRB datasets of interest, it will produce the characteristic
 Islet Regulome Broser plot.
}
